requirements.txt
argon2-cffi==19.1.0
bcrypt==3.1.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Django==2.2.2
django-dynamic-formsets==0.0.8
django-filter==2.2.0
django-filters==0.2.1
django-jalali==3.1.0
django-leaflet==0.24.0
idna==2.8
jdatetime==3.2.0
kavenegar==1.1.2
Pillow==6.0.0
pkg-resources==0.0.0
psycopg2-binary==2.8.3
pycparser==2.19
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
